requirements.txt
beautifulsoup4==4.5.3
Flask==1.0.2
gunicorn==19.9.0
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-language==1.1.1
google-cloud-vision==0.35.2
grpcio==1.18.0
grpcio-tools==1.18.0
requests==2.21.0